VAD/requirements.txt
faiss_cpu==1.8.0
gradio==5.27.0
jieba==0.42.1
numpy==2.2.5
openai==1.76.0
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
rank_bm25==0.2.2
scikit_learn==1.6.1
ultralytics==8.3.51
